Fix Javascipt error
