+++ app/assets/stylesheets/application.css
@@ -92,6 +92,7 @@ nav {
   display: grid;
   grid-auto-flow: column;
   margin-right: 20px;
+  border-bottom: 2px solid var(--primary-color);
 }
 
 .left,
@@ -108,6 +109,8 @@ nav {
 
 .category-link {
   display: list-item;
+  text-transform: uppercase;
+  letter-spacing: 1.5px;
 }
 
 .right {
@@ -144,14 +147,13 @@ footer a {
 .home-best-article-container a {
   display: grid;
   grid-template-rows: 400px 4rem max-content;
-  background-color: var(--primary-color);
-  color: var(--on-primary);
   transition: 0.3s;
   grid-gap: 10px;
   border-radius: 15px;
   box-shadow: var(--box-shadow);
-  border: 1px solid var(--primary-color);
   padding: 30px;
+  color: var(--on-background);
+  background-color: var(--background);
 }
 
 .home-best-article-container a:hover {
@@ -170,14 +172,13 @@ footer a {
 .home-article-container a {
   display: grid;
   grid-template-rows: 350px 30px max-content;
-  background-color: var(--primary-color);
-  color: var(--on-primary);
   transition: 0.3s;
   grid-gap: 8px;
   border-radius: 15px;
   box-shadow: var(--box-shadow);
-  border: 1px solid var(--primary-color);
   padding: 15px;
+  color: var(--on-background);
+  background-color: var(--background);
 }
 
 .home-article-container a:hover {
