+++ app/javascript/packs/application.js
@@ -3,10 +3,11 @@
 // a relevant structure within app/javascript and only use these pack files to reference
 // that code so it'll be compiled.
 
-require("@rails/ujs").start();
+// require("@rails/ujs").start();
 require("turbolinks").start();
 require("@rails/activestorage").start();
 require("channels");
+require("../toggle-mode");
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
@@ -14,57 +15,3 @@ require("channels");
 //
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
-
-// Dark-Light-Mode Toggle
-console.log("reset js");
-var toggleSwitch = document.querySelector('input[type="checkbox"]');
-var nav = document.getElementById("nav");
-var toggleIcon = document.getElementById("toggle-icon");
-
-// Dark Mode Styles
-function darkMode() {
-  console.log("nav", nav);
-  nav.style.backgroundColor = "rgb(0 0 0 / 50%)";
-  // textBox.style.backgroundColor = "rgb(255 255 255 / 50%)";
-  toggleIcon.children[0].textContent = "Dark Mode";
-  toggleIcon.children[1].classList.replace("fa-sun", "fa-moon");
-  // imageMode("dark");
-}
-
-// Light Mode Styles
-function lightMode() {
-  nav.style.backgroundColor = "rgb(255 255 255 / 50%)";
-  // textBox.style.backgroundColor = "rgb(0 0 0 / 50%)";
-  toggleIcon.children[0].textContent = "Light Mode";
-  toggleIcon.children[1].classList.replace("fa-moon", "fa-sun");
-  // imageMode("light");
-}
-
-// Switch Theme Dynamically
-function switchTheme(event) {
-  // console.log(event.target.checked);
-  if (event.target.checked) {
-    document.documentElement.setAttribute("data-theme", "dark");
-    localStorage.setItem("theme", "dark");
-    darkMode();
-  } else {
-    document.documentElement.setAttribute("data-theme", "light");
-    localStorage.setItem("theme", "light");
-    lightMode();
-  }
-}
-
-// Event Listener
-toggleSwitch.addEventListener("change", switchTheme);
-
-// Check Local Storage for Theme
-var currentTheme = localStorage.getItem("theme");
-
-if (currentTheme) {
-  document.documentElement.setAttribute("data-theme", currentTheme);
-
-  if (currentTheme === "dark") {
-    toggleSwitch.checked = true;
-    darkMode();
-  }
-}
