+++ app/views/layouts/application.html.erb
@@ -8,6 +8,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <%= stylesheet_link_tag 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.10.2/css/all.min.css' %>
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
+    <%= stylesheet_link_tag 'https://fonts.googleapis.com/css2?family=Pangolin&display=swap' %>
+    
+
   </head>
 
   <body>
@@ -20,6 +23,5 @@
 
     <%= render "layouts/footer" %>
     <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
-
   </body>
 </html>
